fix(KEN-877): Address PR review - the marker only counts after a value
`# required` written above the assignment instead of after the value parsed
as an ordinary comment and nothing said so. Both readers reported no
marker, findings came back empty, `kendex check --catalog . --strict` exited
0, and the whole suite stayed green. A key an author declared the consumer
must decide was simply never written.

The templates invite it. All nine headers name `# required` inside a
sentence, so an author meets the word in a comment line before they ever
meet it after a value.

Decision 3 raises the severity rather than settling it. While a marker that
fails to parse only meant the key went unwritten, the consumer had a missing
line to notice. Now that an unanswered required key is REPORTED, a marker
that fails to parse means there is no report either: nothing downstream
knows the key was ever marked, so it is invisible rather than merely
unwritten. The two changes compose badly, and this is the condition that
stops them.

The rule is one line of the scan: a comment line whose whole text is the
marker word. Every occurrence in the nine shipped templates is backticked
inside a longer sentence, never a line of its own, so nothing here trips it
and the narrow spelling is the mistake and only the mistake.

Both marker findings move out of `scan` into `marker_alone` and
`marker_after_value`, which puts the two halves of one rule beside each
other: the marker means something after a value and nowhere else, and each
way of getting that wrong reads the same to the loaders as getting it
right.

That took settings_template.rs over the 400-line ratchet, and the seam is
the one the module doc already drew. The types and what a read amounts to
stay in the parent; the line walk that decides what is wrong with a
template, which is the half that grows every time the grammar gains a rule,
moves to settings_template/scan.rs with the three helpers only it calls.

Controls, both red before green. The unit finding at
`settings_template::tests::a_marker_on_its_own_comment_line_is_located`,
which also holds the two things that are NOT the mistake: the word inside a
sentence, and a marker above a table rather than a key. And the same defect
end to end at
`catalog_check::a_marker_on_its_own_comment_line_fails_marketplace_check`,
where `marketplace check` exits nonzero and names the line. Disabling the
condition fails each.

The asymmetry review saw and left to judgement stays as it is: `"K" = ""
# required` reads as marked on the seed side while the strict reader emits
no entry and a separate finding. That is the quoted key's own pre-existing
split — the presence check is deliberately wider than what the loaders read
— and strict catches it, so it is not silent.
